requirements.txt
asgiref==3.3.1
gunicorn==20.1.0
pytz==2021.1
sqlparse==0.4.1
waitress==2.0.0
whitenoise==5.2.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.19.5
matplotlib==3.3.4
Pillow==8.1.0
cv2-tools==2.4.0
scikit-image==0.18.1
opencv-python-headless
tensorflow-cpu
gevent==21.1.2
Jinja2==3.0.1
Flask==2.0.1